Externalize scrolling extensions
